rename ELFunctionManager to ShardRouteRuleExpressionContext
